Auth code restructurized
