+++ src/routes/auth.rs
@@ -1,4 +1,5 @@
 use chrono::{Duration, Utc};
+use rocket::http::{Cookie, Cookies};
 use rocket::State;
 use rocket_contrib::json::{Json, JsonValue};
 
@@ -15,6 +16,7 @@ pub fn login(
     login_data: Json<UserLoginData>,
     c: DBConnection,
     state: State<AppState>,
+    mut cookies: Cookies,
 ) -> JsonValue {
     let user_login_data = login_data.into_inner();
     let result = db::users::login(&c, &user_login_data.username, &user_login_data.password);
@@ -22,10 +24,13 @@ pub fn login(
         Ok(user) => {
             let token = AuthClaims {
                 id: user.id,
-                exp: (Utc::now() + Duration::days(TOKEN_EXP)).timestamp(),
+                exp: (Utc::now() + Duration::minutes(TOKEN_EXP)).timestamp(),
                 iat: Utc::now().timestamp(),
             }
             .create_token(&state.secret);
+
+            cookies.add(Cookie::new("refresh-token", "")); // TODO: provide refresh token here
+
             json!({ "token": token })
         }
         Err(error) => {
