+++ src/routes/mod.rs
@@ -1,3 +1,3 @@
 pub mod articles;
+pub mod auth;
 pub mod catchers;
-pub mod login;
