+++ src/server.rs
@@ -35,7 +35,7 @@ pub fn run() {
                 routes::articles::article_create,
                 routes::articles::article_update,
                 routes::articles::article_archive,
-                routes::login::login,
+                routes::auth::login,
             ],
         )
         .attach(config::AppState::secret_retriever())
